+++ ContentPatcher/Framework/Tokens/ValueProviders/ModConvention/ConventionValueProvider.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
@@ -37,8 +38,11 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.ModConvention
         /// <remarks>This is cached to ensure it never changes outside a context update (even if the mod token is otherwise incorrectly changing without a context update), since that would cause subtle hard-to-troubleshoot bugs where patches don't update correctly in some cases.</remarks>
         public bool IsReady { get; private set; }
 
-        /// <summary>Whether to allow using this token in any value context (e.g. as a number or boolean) without validating ahead of time.</summary>
+        /// <inheritdoc />
         public bool BypassesContextValidation => false;
+
+        /// <inheritdoc />
+        public Func<string, string>? NormalizeValue => this.Provider.NormalizeValue;
 
 
         /*********
@@ -101,12 +105,6 @@ namespace ContentPatcher.Framework.Tokens.ValueProviders.ModConvention
         public IEnumerable<string> GetValues(IInputArguments input)
         {
             return this.Provider.GetValues(this.ToApiInput(input));
-        }
-
-        /// <inheritdoc />
-        public string? NormalizeValue(string? value)
-        {
-            return this.Provider.NormalizeValue(value);
         }
 
         /// <inheritdoc />